Uniform Block and LightManager Refactor
Moved more Uniform Block logic into UniformBlockInfo so LightManager and Renderer no longer need to know about it. (This is also a step in removing the use of friend in my architecture.

LightManager's shadow texture is now a member variable instead of the hard-coded 31 I used because I was coding under a severe time crunch... At least I fixed it.

Added a missing lib file.